\name{addmod2}
\alias{addmod2}
\title{addmod2}
\usage{
addmod2(x,a)
}
\arguments{
  \item{x}{ integer }

  \item{a}{ integer  }
}

\value{
 Integer resulting from the bit-wise addition of two number modulus 2
}

\description{
Takes two integers and adds their bits modulus two. The resulting string of bits represents an integer, that value of which is the output. 
}

\examples{
 addmod2(5,5)
 addmod2(1,2)
}
