\name{bra}
\alias{bra}
\title{bra}
\usage{
bra(...)
}
\arguments{
  \item{...}{ Variable number of numbers representing the amplitudes of the bra   }
}

\value{
Row vector containing normalized amplitudes of a bra
}

\description{
Returns a normalized bra (row vector) 
}

\examples{
 bra(1,0,1,2)
}
