\name{ket}
\alias{ket}
\title{ket}
\usage{
ket(...)
}
\arguments{
  \item{...}{ Variable number of numbers representing the amplitudes of the ket   }
}

\value{
Column vector containing normalized amplitudes of a ket
}

\description{
Returns a normalized ket (column vector) 
}

\examples{
 ket(1,0,1,2)
}
