\name{mm}
\alias{mm}
\title{mm}
\usage{
mm(...)
}
\arguments{
  \item{...}{ Variable number of numbers representing the values in the matrix   }
}

\value{
Matrix containing the values of the inputs
}

\description{
Returns a matrix containing the specified elements. Values are input column-wise. Used for convenient shorthand creation of matrices
}

\examples{
 mm(1,0,1,2)
}
