\name{QAOA_example}
\alias{QAOA_example}
\title{QAOA_example}
\usage{
QAOA_example(case=1)
}
\arguments{
  \item{case}{Integer specifying case to demonstrate. Currently only two, 1 (small) and 2 (medium)}
}

\value{
No value
}

\description{
Runs an example of QAOA
}

\examples{
 \dontrun{
 QAOA_example(1)
 QAOA_example(2)
 }
}
